Inverted the way AudioSpectrumSelector buttons are made so that they correspond to the equalizer band they appear beneath.  They are still referenced as 0-6 (or really 0 to NUM_FREQS-1) but are ordered from 6 to 0.  For conventions sake, the lowest bass frequencies (Band 7) appears first but is labeled as band 0.
Constrained maximum equalizer rectangle height to 255, same height as the sliders

Made prettier sliders

Pushed the text over by 5 px in the Generic Button class.  Still thinking about making prettier buttons

Made balloons highlighted in black with an alpha of 100.  This is nice, but what we need is a deselect button or just a way of rendering them the default way when you click on white space.  At the moment the last group of balloons remain highlighted and you can't deselect.  My understanding of your code isn't good enough to figure that out.

The code for highlighting independent groups of balloons appears to still be there but doesn't seem to be working?

Finally, the mouse click draws a highlighting rectangle everywhere.  This would be nice to keep limited to the balloon area - kind of weird when it does it as you are clicking and dragging a slider button.

How's that for a wish list?